+++ src/interpreter/Interpreter.ts
@@ -3,7 +3,6 @@
  * Copyright (c) 2020 Handsoncode. All rights reserved.
  */
 
-import GPU from 'gpu.js'
 import {AbsoluteCellRange, AbsoluteColumnRange, AbsoluteRowRange} from '../AbsoluteCellRange'
 import {CellError, ErrorType, invalidSimpleCellAddress, SimpleCellAddress} from '../Cell'
 import {Config} from '../Config'
@@ -28,13 +27,14 @@ import {
   EmptyValue,
   getRawValue,
   InternalNoErrorScalarValue,
-  InterpreterValue,
   isExtendedNumber,
 } from './InterpreterValue'
+import {InterpreterValue} from './InterpreterValue'
+import type {GPU} from 'gpu.js'
 import {SimpleRangeValue} from './SimpleRangeValue'
 
 export class Interpreter {
-  private gpu?: GPU.GPU
+  private gpu?: GPU
   public readonly arithmeticHelper: ArithmeticHelper
   public readonly criterionBuilder: CriterionBuilder
 
@@ -65,6 +65,7 @@ export class Interpreter {
     }
     return wrapperForAddress(val, formulaAddress)
   }
+
   /**
    * Calculates cell value from formula abstract syntax tree
    *
@@ -209,7 +210,7 @@ export class Interpreter {
         }
       }
       case AstNodeType.FUNCTION_CALL: {
-        if(this.config.licenseKeyValidityState !== LicenseKeyValidityState.VALID && !FunctionRegistry.functionIsProtected(ast.procedureName)) {
+        if (this.config.licenseKeyValidityState !== LicenseKeyValidityState.VALID && !FunctionRegistry.functionIsProtected(ast.procedureName)) {
           return new CellError(ErrorType.LIC, ErrorMessage.LicenseKey(this.config.licenseKeyValidityState))
         }
         const pluginEntry = this.functionRegistry.getFunction(ast.procedureName)
@@ -273,16 +274,16 @@ export class Interpreter {
     }
   }
 
-  public getGpuInstance(): GPU.GPU {
+  public getGpuInstance(): GPU {
     const mode = this.config.gpuMode
+    const gpujs = this.config.gpujs
 
-    if (mode === 'fallback') {
-      throw Error('Cannot instantiate GPU.js when fallback mode is active.')
+    if (gpujs === undefined) {
+      throw Error('Cannot instantiate GPU.js. Constructor not provided.')
     }
 
     if (!this.gpu) {
-      const GPUConstructor = GPU.GPU || GPU
-      this.gpu = new GPUConstructor({ mode })
+      this.gpu = new gpujs({mode})
     }
 
     return this.gpu
@@ -314,7 +315,7 @@ function passErrors(left: InterpreterValue, right: InterpreterValue): Maybe<Cell
 }
 
 function wrapperUnary<T extends InterpreterValue>(op: (a: T) => InterpreterValue, a: T | CellError): InterpreterValue {
-  if(a instanceof CellError) {
+  if (a instanceof CellError) {
     return a
   } else {
     return op(a)
@@ -322,9 +323,9 @@ function wrapperUnary<T extends InterpreterValue>(op: (a: T) => InterpreterValue
 }
 
 function wrapperBinary<T extends InterpreterValue>(op: (a: T, b: T) => InterpreterValue, a: T | CellError, b: T | CellError): InterpreterValue {
-  if(a instanceof CellError) {
+  if (a instanceof CellError) {
     return a
-  } else if(b instanceof CellError) {
+  } else if (b instanceof CellError) {
     return b
   } else {
     return op(a, b)
@@ -332,7 +333,7 @@ function wrapperBinary<T extends InterpreterValue>(op: (a: T, b: T) => Interpret
 }
 
 function wrapperForAddress(val: InterpreterValue, adr: SimpleCellAddress): InterpreterValue {
-  if(val instanceof CellError) {
+  if (val instanceof CellError) {
     return val.attachAddress(adr)
   }
   return val